requirements.txt
numpy==1.26.3
pickle-mixin==1.0.2
streamlit-option-menu==0.3.2
scikit-learn==1.0.1
protobuf==3.20.1
pandas==2.1.4
numpy==1.23.5
scikit-learn==1.2.2
Pillow==9.4.0
joblib==1.3.2
streamlit==1.30.0
streamlit
pandas
numpy
scikit-learn
scikit-learn==1.2.2
nltk


